Move platform-specific syscalls modules into separate files
