+++ src/debug/low_level_debug.rs
@@ -2,7 +2,7 @@
 //! useful for toolchain issues. The capsule is documented at:
 //! https://github.com/tock/tock/blob/master/doc/syscalls/00008_low_level_debug.md
 
-use crate::syscalls::{command, command1_insecure};
+use crate::syscalls;
 
 const DRIVER_NUMBER: usize = 8;
 
@@ -17,7 +17,7 @@ mod command_nr {
 #[inline(always)] // Improve reliability for relocation issues
 pub fn low_level_status_code(code: usize) {
     unsafe {
-        command1_insecure(DRIVER_NUMBER, command_nr::ALERT_CODE, code);
+        syscalls::raw::command1(DRIVER_NUMBER, command_nr::ALERT_CODE, code);
     }
 }
 
@@ -26,7 +26,7 @@ pub fn low_level_status_code(code: usize) {
 #[inline(always)] // Improve reliability for relocation issues
 pub fn low_level_print1(value: usize) {
     unsafe {
-        command1_insecure(DRIVER_NUMBER, command_nr::PRINT1, value);
+        syscalls::raw::command1(DRIVER_NUMBER, command_nr::PRINT1, value);
     }
 }
 
@@ -35,6 +35,6 @@ pub fn low_level_print1(value: usize) {
 #[inline(always)] // Improve reliability for relocation issues
 pub fn low_level_print2(value1: usize, value2: usize) {
     unsafe {
-        command(DRIVER_NUMBER, command_nr::PRINT2, value1, value2);
+        syscalls::raw::command(DRIVER_NUMBER, command_nr::PRINT2, value1, value2);
     }
 }
