+++ src/memop.rs
@@ -8,11 +8,11 @@ use core::slice;
 //
 // Alternate allocator implementations may still find this useful in the future.
 pub unsafe fn set_brk(ptr: *const u8) -> bool {
-    syscalls::memop(0, ptr as usize) == 0
+    syscalls::raw::memop(0, ptr as usize) == 0
 }
 
 pub unsafe fn increment_brk(increment: usize) -> Option<*const u8> {
-    let result = syscalls::memop(1, increment);
+    let result = syscalls::raw::memop(1, increment);
     if result >= 0 {
         Some(result as *const u8)
     } else {
@@ -21,36 +21,36 @@ pub unsafe fn increment_brk(increment: usize) -> Option<*const u8> {
 }
 
 pub fn get_brk() -> *const u8 {
-    unsafe { syscalls::memop(1, 0) as *const u8 }
+    unsafe { syscalls::raw::memop(1, 0) as *const u8 }
 }
 
 pub fn get_mem_start() -> *const u8 {
-    unsafe { syscalls::memop(2, 0) as *const u8 }
+    unsafe { syscalls::raw::memop(2, 0) as *const u8 }
 }
 
 pub fn get_mem_end() -> *const u8 {
-    unsafe { syscalls::memop(3, 0) as *const u8 }
+    unsafe { syscalls::raw::memop(3, 0) as *const u8 }
 }
 
 pub fn get_flash_start() -> *const u8 {
-    unsafe { syscalls::memop(4, 0) as *const u8 }
+    unsafe { syscalls::raw::memop(4, 0) as *const u8 }
 }
 
 pub fn get_flash_end() -> *const u8 {
-    unsafe { syscalls::memop(5, 0) as *const u8 }
+    unsafe { syscalls::raw::memop(5, 0) as *const u8 }
 }
 
 pub fn get_grant_start() -> *const u8 {
-    unsafe { syscalls::memop(6, 0) as *const u8 }
+    unsafe { syscalls::raw::memop(6, 0) as *const u8 }
 }
 
 pub fn get_flash_regions_count() -> usize {
-    unsafe { syscalls::memop(7, 0) as usize }
+    unsafe { syscalls::raw::memop(7, 0) as usize }
 }
 
 pub fn get_flash_region_start(i: usize) -> Option<*const u8> {
     if i < get_flash_regions_count() {
-        Some(unsafe { syscalls::memop(8, i) as *const u8 })
+        Some(unsafe { syscalls::raw::memop(8, i) as *const u8 })
     } else {
         None
     }
@@ -58,7 +58,7 @@ pub fn get_flash_region_start(i: usize) -> Option<*const u8> {
 
 pub fn get_flash_region_end(i: usize) -> Option<*const u8> {
     if i < get_flash_regions_count() {
-        Some(unsafe { syscalls::memop(9, i) as *const u8 })
+        Some(unsafe { syscalls::raw::memop(9, i) as *const u8 })
     } else {
         None
     }
@@ -68,8 +68,8 @@ pub fn get_flash_region_end(i: usize) -> Option<*const u8> {
 // flash regions during the application's lifetime.
 pub fn get_flash_region(i: usize) -> Option<&'static [u8]> {
     if i < get_flash_regions_count() {
-        let start = unsafe { syscalls::memop(8, i) as *const u8 };
-        let end = unsafe { syscalls::memop(9, i) as *const u8 };
+        let start = unsafe { syscalls::raw::memop(8, i) as *const u8 };
+        let end = unsafe { syscalls::raw::memop(9, i) as *const u8 };
         // This assumes that the kernel sends consistent results, i.e. start <= end.
         let len = unsafe { end.offset_from(start) } as usize;
         Some(unsafe { slice::from_raw_parts(start, len) })
@@ -84,9 +84,9 @@ pub fn get_flash_region(i: usize) -> Option<&'static [u8]> {
 //
 // Alternate allocator implementations may still find this useful in the future.
 pub unsafe fn set_stack_top(ptr: *const u8) {
-    let _ = syscalls::memop(10, ptr as usize);
+    let _ = syscalls::raw::memop(10, ptr as usize);
 }
 
 pub unsafe fn set_heap_start(ptr: *const u8) {
-    let _ = syscalls::memop(11, ptr as usize);
+    let _ = syscalls::raw::memop(11, ptr as usize);
 }
