+++ src/sensors/ninedof.rs
@@ -65,7 +65,7 @@ impl Ninedof {
 }
 
 pub unsafe fn subscribe(cb: extern "C" fn(usize, usize, usize, usize), ud: usize) {
-    syscalls::subscribe_ptr(DRIVER_NUM, 0, cb as *const _, ud);
+    syscalls::subscribe_fn(DRIVER_NUM, 0, cb, ud);
 }
 
 pub unsafe fn start_accel_reading() {
